:adhesive_bandage: fix(janelasHorario e Agendamento): implementando a logica de janelas de horario no agendamento

+++ src/main/java/com/fatec/itu/agendasalas/dto/agendamentosDTO/AgendamentoAulaResponseDTO.java
@@ -7,7 +7,7 @@ public record AgendamentoAulaResponseDTO(
     Long id,
     String nomeUsuario,
     String nomeSala,
-    Integer disciplinaId,
+    Long disciplinaId,
     String nomeDisciplina,
     String semestre,
     String curso,
@@ -18,6 +18,7 @@ public record AgendamentoAulaResponseDTO(
     LocalTime horaInicio,
     LocalTime horaFim,
     String tipo) {
+
     
    
 }
